lets see-real shit

+++ src/main/java/com/example/demo/Repository/FormRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.entity.FormEntity;
 
-public interface FormRepo extends JpaRepository<FormEntity, Integer> {
+public interface FormRepo extends JpaRepository<FormEntity, Long> {
 
 }
